requirements.txt
numpy
opencv-python
pandas
matplotlib
pyparsing
g2o-python